systeme d auth avec better auth

+++ app/api/recruteur/offres/[id]/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import prisma from "@/lib/prisma";
-import { getAuthenticatedUser } from "@/lib/auth-utils";
+
 import { auth } from "@/lib/auth";
 // import { cacheUtils, CACHE_KEYS, CACHE_TTL } from "@/lib/redis";
 
@@ -11,7 +11,11 @@ export async function PUT(
   try {
     const id = (await params).id;
     const body = await req.json();
-    console.log("Received skills in API:", body.skills);
+    const session = await auth.api.getSession({ headers: req.headers });
+    if (!session) {
+      return NextResponse.json({ error: "Non autorisé" }, { status: 401 });
+    }
+
     const {
       title,
       description,
@@ -31,6 +35,20 @@ export async function PUT(
       etat,
       // recruteurId,
     } = body;
+
+    const recruteur = await prisma.recruteur.findUnique({
+      where: { userId: session.user.id },
+    });
+
+    const collaborateur = await prisma.collaborateur.findFirst({
+      where: { userId: session.user.id },
+      include: {
+        recruteur: true,
+      },
+    });
+    if (!recruteur && !collaborateur) {
+      return NextResponse.json({ error: "Non autorisé" }, { status: 401 });
+    }
 
     // First, delete existing competences for this job offer
     await prisma.jobOfferCompetence.deleteMany({
@@ -98,10 +116,17 @@ export async function GET(
         where: { userId: session.user.id },
       });
 
+      const collaborateur = await prisma.collaborateur.findFirst({
+        where: { userId: session.user.id },
+        include: {
+          recruteur: true,
+        },
+      });
+
       const offer = await prisma.jobOffer.findFirst({
         where: {
           id: Number(offerId),
-          recruteurId: recruteurId?.id,
+          recruteurId: recruteurId?.id || collaborateur?.recruteur?.id || "",
         },
         include: {
           kanbanColumns: { orderBy: { order: "asc" } },
@@ -193,6 +218,25 @@ export async function DELETE(
   try {
     const id = (await params).id;
 
+    const session = await auth.api.getSession({ headers: req.headers });
+    if (!session) {
+      return NextResponse.json({ error: "Non autorisé" }, { status: 401 });
+    }
+
+    const collaborateur = await prisma.collaborateur.findFirst({
+      where: { userId: session.user.id },
+      include: {
+        recruteur: true,
+      },
+    });
+
+    const recruteur = await prisma.recruteur.findUnique({
+      where: { userId: session.user.id },
+    });
+    if (!recruteur && !collaborateur) {
+      return NextResponse.json({ error: "Non autorisé" }, { status: 401 });
+    }
+
     // Supprimer d'abord les colonnes du kanban associées
     await prisma.kanbanColumn.deleteMany({
       where: {
@@ -212,7 +256,6 @@ export async function DELETE(
       { status: 200 }
     );
   } catch (err) {
-    console.log(err);
     return NextResponse.json(
       {
         success: false,
